Added Mockito and unit tests for each of the controllers.

+++ pet-clinic-visits-service/src/main/java/com/rje/petclinic/repositories/VisitRepository.java
@@ -9,6 +9,6 @@ public interface VisitRepository extends CrudRepository<Visit, Long> {
 
     List<Visit> findByPetId(Long petId);
 
-    List<Visit> findByPetIdIn(List<Integer> petIds);
+    List<Visit> findByPetIdIn(List<Long> petIds);
 
 }
